Add project description and MIT license
Visualize the status of multiple support requests for a WordPress
plugin. This tools sends HTTP requests to the latest twenty pages of the
support page of the specified plugin, finds how many of the tickets are
marked as resolved, and shows which ones are missing. Many people in the
WordPress community sees the number of resolved tickets per month as a
sign of responsiveness and promptness, other use this as one of the main
reasons to install or not a plugin.