Add main function with intended functionality
The program will send multiple HTTP GET requests to the WordPress forum
associated to the specified plugin, the requests are fired concurrently
and printed unordered. Pages will too many unresolved tickets are marked
with a warning or error symbol while the pages with many resolved tickets
are marked with a green / successful symbol.